third_party/blink/web_tests/platform/generic/editing/deleting/delete-non-static-position-expected.txt

PASS document.querySelectorAll("p.li").length is 5
PASS successfullyParsed is true

TEST COMPLETE
